Fix red team results data grid tests
