+++ src/app/src/App.tsx
@@ -15,14 +15,12 @@ import { useTelemetry } from './hooks/useTelemetry';
 import DatasetsPage from './pages/datasets/page';
 import EvalPage from './pages/eval/page';
 import EvalCreatorPage from './pages/eval-creator/page';
-import EvalsIndexPage from './pages/evals/page';
-import HistoryPage from './pages/history/page';
+import ResultsPage from './pages/results/page';
 import LauncherPage from './pages/launcher/page';
 import LoginPage from './pages/login';
 import ModelAuditPage from './pages/model-audit/page';
 import NotFoundPage from './pages/NotFoundPage';
 import PromptsPage from './pages/prompts/page';
-import ReportPage from './pages/redteam/report/page';
 import RedteamSetupPage from './pages/redteam/setup/page';
 
 const basename = import.meta.env.VITE_PUBLIC_BASENAME || '';
@@ -57,21 +55,22 @@ const router = createBrowserRouter(
           />
           <Route path="/datasets" element={<DatasetsPage />} />
           <Route path="/eval" element={<EvalPage />} />
-          <Route path="/evals" element={<EvalsIndexPage />} />
+          <Route path="/evals" element={<Navigate to="/results" replace />} />
+          <Route path="/results" element={<ResultsPage />} />
           <Route path="/eval/:evalId" element={<EvalPage />} />
 
-          {/* Redirect legacy /progress route to /history (since v0.104.5) */}
-          <Route path="/progress" element={<Navigate to="/history" replace />} />
-          <Route path="/history" element={<HistoryPage />} />
+          {/* Redirect legacy /progress route to /results?tab=evals (since v0.104.5) */}
+          <Route path="/progress" element={<Navigate to="/results?tab=evals" replace />} />
+          <Route path="/history" element={<Navigate to="/results?tab=evals" replace />} />
 
           <Route path="/prompts" element={<PromptsPage />} />
           <Route path="/model-audit" element={<ModelAuditPage />} />
           <Route path="/redteam" element={<Navigate to="/redteam/setup" replace />} />
           <Route path="/redteam/setup" element={<RedteamSetupPage />} />
 
-          {/* Redirect legacy /report route to /reports (since v0.118.2) */}
-          <Route path="/report" element={<Navigate to="/reports" replace />} />
-          <Route path="/reports" element={<ReportPage />} />
+          {/* Redirect legacy /report route to /results?tab=redteams (since v0.118.2) */}
+          <Route path="/report" element={<Navigate to="/results?tab=redteams" replace />} />
+          <Route path="/reports" element={<Navigate to="/results?tab=redteams" replace />} />
           <Route path="/setup" element={<EvalCreatorPage />} />
           <Route path="/login" element={<LoginPage />} />
           <Route path="*" element={<NotFoundPage />} />
